requirements.txt
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
databases==0.7.0
dill==0.3.7
fastapi==0.101.1
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
h11==0.14.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector-python==8.1.0
platformdirs==3.10.0
protobuf==4.21.12
pydantic==2.1.1
pydantic_core==2.4.0
pylint==2.17.5
pyserial==3.5
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
tomlkit==0.12.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
Werkzeug==2.3.7
wrapt==1.15.0
